Modify : 카테고리 기능 오류 수정
